Update ietf-tpm-remote-attestation.yang
The identities for TPM 1.2 and 2.0 already existed in TCG algs.  Eliminated them here.   Added a leaf which enables an Attester to limit which TPM1.2 PCRs might be acquired.   Added some initial XPATH to limit which PCRs can be extracted as part of an RPC.   They still need work, and am hoping the YANG doctors can provide the syntax for this error check.